fix: Fix bug in how the test function builder generates try that exit early
Previously, PythonBytecodeInstruction was mutable. This meant that
the test function builder can pass the except builder a dummy
instruction for the try goto when its exit early.

PythonBytecodeInstruction is now immutable, which means
to update the instruction, you must replace it in the list.
This causes issues when the dummy instruction is used, since
the dummy instruction will replace a real instruction.

The test function builder will now pass null instead of
a dummy instruction, which except builder need to check
before updating the list.