fix: Load globals that are subclasses or instances of a class being compiled lazily
Previously, the entire global dict was loaded whenever a function was
compiled. However, this causes issues if there is a global that is an
instance or subclass of a class being compiled, and the class being
compiled reference that global.

Now, when a class is being compiled, instances and subclasses of that
class is excluded and loaded at runtime.