fix: Use Number.doubleValue() instead of float(Number) for deserialization
When testing VRP, it was discovered that occasionally, the location,
represented as a pair of floats, was converted incorrectly.

It appears JPype's float conversion for Number is inconsistent, so
we use Number.doubleValue() instead, which seems to consistency
convert correctly.

Ideally, `test_float.test_use_64_bits` or the other float tests
would have catched it, but somehow, those tests never failed;
the exact condition for incorrect conversion is unknown.